ingestion/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "thrn-ingest"
version = "0.1.0"
description = "OpenAlex ingestion pipeline for the Tourism & Hospitality Research Navigator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
dependencies = [
    "psycopg[binary,pool]>=3.2",
    "requests>=2.32",
    "tenacity>=9.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "typesense>=0.21",
    "orjson>=3.10",
]

[project.scripts]
thrn = "thrn.__main__:app"

[tool.setuptools.packages.find]
where = ["src", "cli"]

[tool.setuptools.package-dir]
"" = "src"
"thrn" = "cli/thrn"
